requirements.txt
pandas
datetime
azure-ai-evaluation
aiohttp
transformers==4.44.2
peft~=0.12
accelerate~=0.33
trl==0.10.1
einops==0.8.0
datasets==2.21.0
wandb==0.17.8
mlflow==2.16.0
azureml-mlflow==1.57.0
azureml-sdk==1.57.0
torchvision==0.19.0
torch==2.4.0
azureml-core==1.58.0
azureml-dataset-runtime==1.58.0
azureml-defaults==1.58.0
azure-ml==0.0.1
azure-ml-component==0.9.18.post2
azureml-contrib-services==1.58.0
azureml-automl-common-tools==1.58.0
torch-tb-profiler==0.4.3
scipy
azure-identity
packaging==24.1
timm==1.0.11